requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
openai==1.12.0
pandas==2.2.0
python-decouple==3.8
